begin reworking/merging system of converter<t> and type_wrap<t>
